pyproject.toml
[tool.poetry]
name = "nautobot-eox-notices"
version = "0.1.0"
description = "Tracks EoX Notices for Nautobot objects."
authors = ["Mikhail Yohman <mikhail.yohman@gmail.com>"]

license = "Apache-2.0"

readme = "README.md"
homepage = "https://github.com/fragmentedpacket/nautobot-eox-notices"
repository = "https://github.com/fragmentedpacket/nautobot-eox-notices"
keywords = ["nautobot", "nautobot-plugin"]
include = [

    "LICENSE",

    "README.md",
]
packages = [
    { include = "eox_notices" },
]

[tool.poetry.dependencies]
python = "^3.6"

[tool.poetry.dev-dependencies]
invoke = "^1.4.1"
black = "^19.10b0"
yamllint = "^1.23.0"
bandit = "^1.6.2"
pylint = "^2.5.2"
pylint-django = "^2.0.15"
pydocstyle = "^5.0.2"
flake8 = "^3.8.3"
pytest-django = "^4.1.0"

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins="pylint_django"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx="^(_|test_|Meta$)"

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long,
    bad-continuation,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_paths = "./"
testpaths = "tests/"
adopts = "-vv"
